Variabilise the name of the Tsuru PPA
So that we can point it at our own PPA which contains copies of the packages
at the versions we want to use. This is to workaround a limitation of
Launchpad PPAs which only allow you to download the latest version of a
package.

I can't see any nice way (with Ansible) to remove the old repo if you've
specified a custom one, because another playbook (tsuru_api/gandalf/hipache)
may still configure the original repo and both tasks would fight (changing
it back and forth in every run).

So if we use this, we'll also want to either a) remove the old repo as a
custom task outside of these playbooks, or b) specify the package versions
we want so that the old repo is ignored if it presents a later version.